Refactored backend code
Renamed server.js and index.js appropriately, removed unnecessary Database classes and isolated mongo connection to server.js. Fixed small typo in .env and refactored mongodb session to use the same DB URI.
