+++ server/.env
@@ -1,5 +1,5 @@
 JUSTONTIME_DB_URI=mongodb://admin:bavys2022@127.0.0.1:27017/justontimedb?directConnection=true
-JUSTONTIME_DB_URI_LOCAL=ADD_YOUR_DB_URI
+JUSTONTIME_DB_URI_LOCAL=mongodb+srv://sanan:i7UXJXuy1WNnxQc1@cluster0.bxpc4.mongodb.net/test
 JUSTONTIME_DB=justontimedb
-SESSION_SECCRET_KEY=bavys2022
+SESSION_SECRET_KEY=bavys2022
 PORT=3000
