--- server/api/databse/DBConnect.js
@@ -1,14 +1,0 @@
-import mongoose from 'mongoose'
-import { Customer } from '../models/customer.model.js'; 
-import { EventOrganizer } from '../models/eventOrganizer.model.js';
-import { Event } from '../models/event.model.js';
-import dotenv from 'dotenv';
-dotenv.config();
-
-
-async function connectDB() {
-    var uri = process.env.JUSTONTIME_DB_URI_LOCAL; //process.env.JUSTONTIME_DB_URI (use this on production)
-    await mongoose.connect(uri);
-}
-
-export { connectDB }
