--- server/api/databse/DBReset.js
@@ -1,14 +1,0 @@
-import mongoose from 'mongoose'
-import { Customer } from '../models/customer.model.js'; 
-import { EventOrganizer } from '../models/eventOrganizer.model.js';
-import { Event } from '../models/event.model.js';
-import dotenv from 'dotenv';
-dotenv.config();
-
-
-async function removeAllData() {
-    Customer.remove({}, () => {});
-    EventOrganizer.remove({}, () => {});
-}
-
-export { removeAllData }
