--- server/index.js
@@ -1,25 +1,0 @@
-import app from './server.js';
-import mongodb from 'mongodb';
-import dotenv from 'dotenv';
-dotenv.config();
-
-const MongoClient = mongodb.MongoClient;
-const port = process.env.PORT || 3000;
-
-MongoClient.connect(
-    process.env.JUSTONTIME_DB_URI_LOCAL, //REPLACE WITH JUSTONTIME_DB_URI FOR PRODUCTION
-    {
-        maxPoolSize:50,
-        wtimeoutMS:2500,
-        useNewUrlParser:true
-    }
-)
-.catch(err => {
-    console.error(err.stack)
-    process.exit(1)
-})
-.then(async client => {
-    app.listen(port, () => {
-        console.log('listening on port '+port)
-    })
-})
